requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
multidict==5.1.0
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.16.0
pytz==2021.1
requests==2.25.1
six==1.15.0
twitchio==1.2.2
typing-extensions==3.7.4.3
urllib3==1.26.4
websockets==8.1
yarl==1.6.3
